backend/requirements.txt
fastapi
uvicorn
python-dotenv
openssa
pydantic
sse-starlette
